sam-app/mqtt/requirements.txt
requests
PyJWT
firebase-admin